+++ src/test/create-user-test.ts
@@ -1,32 +1,18 @@
 import bcrypt from 'bcrypt';
-import axios, { AxiosResponse } from 'axios';
+import { AxiosResponse } from 'axios';
 import { expect } from 'chai';
-import { serverUrl } from '../setup-server';
 import { AppDataSource } from '../data-source.js';
 import { User } from '../entity/User.js';
 import { Repository } from 'typeorm';
 import { tokenCreation } from '../schema/resolvers';
+import { Address } from '../entity/Address';
+import { postQuery } from './postQuery.js';
 
 export interface CreateUserInputData {
   name: string;
   email: string;
   birthDate: string;
   password: string;
-}
-
-export async function postQuery(query: string, variables?, token?: string): Promise<AxiosResponse> {
-  return axios.post(
-    `${serverUrl}graphql`,
-    {
-      query: query,
-      variables: variables,
-    },
-    {
-      headers: {
-        Authorization: token,
-      },
-    },
-  );
 }
 
 async function createUser(inputUser: CreateUserInputData, token?: string): Promise<AxiosResponse> {
@@ -82,10 +68,13 @@ async function checksInputAndStoredUser(inputUser: CreateUserInputData) {
 
 describe('createUser mutation', () => {
   let userRepository: Repository<User>;
+  let addressRepository: Repository<Address>;
   let token: string;
 
   beforeEach(async () => {
+    addressRepository = AppDataSource.getRepository(Address);
     userRepository = AppDataSource.getRepository(User);
+    await addressRepository.delete({});
     await userRepository.delete({});
 
     token = tokenCreation(1);
